pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "numpy>=2.2.3",
    "openai>=1.102.0",
    "pandas>=2.2.3",
    "pillow-heif>=1.1.0",
    "pillow>=11.1.0",
    "plotly>=6.0.0",
    "psycopg2-binary>=2.9.10",
    "scipy>=1.15.2",
    "setuptools>=80.9.0",
    "sqlalchemy>=2.0.40",
    "streamlit>=1.43.0",
    "werkzeug>=3.1.3",
    "imageio>=2.37.0",
    "opencv-python>=4.12.0.88",
    "moviepy>=2.2.1",
]
